divide Lines.py into 2 separate files
